Changes to migrate to UE 5.2 (#197)
* ue5 required changes to SpearSim project, CoreUtils plugin

* update to use latest include order

* rebuild navmesh

* modify SimulationController plugin to ignore UrdfBot and OpenBot.

* disable some ray-tracing params in CameraSensor.cpp as they are no longer present in UE5

* disable physx usage in SonarSensor.cpp as it's no longer supported.

* float is now double precision. Avoid narrowing conversion in SphereAgent.cpp

* fix RecastNavMesh's CellSize and CellHeight deprecation warning

* add comments for  bEnableUndefinedIdentifierWarnings = false;

+++ cpp/unreal_plugins/SimulationController/Source/SimulationController/SphereAgent.cpp
@@ -275,7 +275,7 @@ std::map<std::string, std::vector<uint8_t>> SphereAgent::getObservation() const
         FVector linear_velocity = static_mesh_component_->GetPhysicsLinearVelocity();
         float yaw = rotation_.Yaw;
 
-        observation["compass"] = Std::reinterpret_as<uint8_t>(std::vector<float>{
+        observation["compass"] = Std::reinterpret_as<uint8_t>(std::vector<double>{
             Config::get<float>("SIMULATION_CONTROLLER.SPHERE_AGENT.COMPASS.OFFSET_TO_GOAL_SCALE") * sphere_to_goal.X,
             Config::get<float>("SIMULATION_CONTROLLER.SPHERE_AGENT.COMPASS.OFFSET_TO_GOAL_SCALE") * sphere_to_goal.Y,
             Config::get<float>("SIMULATION_CONTROLLER.SPHERE_AGENT.COMPASS.LINEAR_VELOCITY_SCALE") * linear_velocity.X,
